Fixing resizing errors
Managed to fix it for you! There is one problem though, now every time the Rectangle fails to display, it will simply stop displaying, regardless of what the issue is, I will fix that in the IRC library by intentionally throwing errors on specific issues, but I didn't feel like doing it here. The Display method returns false so the person coding can detect if it failed, but won't be able to figure out the reason.